fix getHeader to be case-insensitive
www.campman.com returns the  "X-BC-ApiLimit-Remaining" header in all lower-case, like "x-bc-apilimit-remaining".   The getRequestsRemaining can not find the header, and returns 0 as the limit every time.  I patched the getHeader function so it will look for the header key using a case-insensitive search, if it can't find an exact match using the hash lookup.  Now, getRequestsRemaining returns the expected number of API requests remaining.  Note that RFC2616 states that field names in the HTTP header are case-insensitive, so this is the correct thing to do.