Fix getHeader to be case-insensitive
On certain BigCommerce stores, getHeader fails because the X-BC-ApiLimit-Remaining header is all lower-case in the HTTP response.  This causes getRequestsRemaining to always return 0 and puts the code into an infinite wait for more requests.  Instead, change getHeader to do a case-insensitive search for the header if the exact case is not found.  Please merge this into the trunk.